agents/requirements.txt
jsonschema==4.25.0
python-dotenv==1.1.1
websockets==15.0.1
langchain-google-genai==2.1.8
langgraph==0.5.3